feat : handleJoin
